ac: fix return type for llvm.amdgcn.frexp.exp.i32.64
This fixes the following piglit with RadeonSI
tests/spec/arb_gpu_shader_fp64/execution/built-in-functions/fs-frexp-dvec4.shader_test

+++ src/amd/common/ac_llvm_build.c
@@ -3942,7 +3942,7 @@ ac_build_frexp_exp(struct ac_llvm_context *ctx, LLVMValueRef src0,
 		type = ctx->i32;
 	} else {
 		intr = "llvm.amdgcn.frexp.exp.i32.f64";
-		type = ctx->i64;
+		type = ctx->i32;
 	}
 
 	LLVMValueRef params[] = {
